loculus-silo/pyproject.toml
[project]
name = "silo-import"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "zstandard>=0.22,<0.26",
    "requests>=2.31,<3.0",
    "orjsonl>=0.3,<2.0",
    "PyYAML>=6.0,<7.0",
]

[project.scripts]
silo-import = "silo_import.__main__:main"

[project.optional-dependencies]
test = ["pytest", "mypy",
    "ruff",]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/silo_import"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q"]
